+++ vector/src/main/java/im/vector/activity/VectorRoomActivity.java
@@ -1419,6 +1419,10 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
         if (mReadMarkerManager != null) {
             mReadMarkerManager.onScrollStateChanged(scrollState);
         }
+
+        if (mNotificationsArea != null) {
+            mNotificationsArea.setScrollState(scrollState);
+        }
     }
 
     @Override
@@ -3130,7 +3134,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
                 // hide the action bar header view and reset the arrow image (arrow reset to down)
                 mActionBarCustomArrowImageView.setImageResource(R.drawable.ic_arrow_drop_down_white);
                 mRoomHeaderView.setVisibility(View.GONE);
-                toolbar.setBackgroundColor(ThemeUtils.INSTANCE.getColor(this, R.attr.vctr_primary_color));
+                toolbar.setBackgroundColor(ThemeUtils.INSTANCE.getColor(this, R.attr.colorPrimary));
             }
         }
     }