+++ vector/src/main/java/im/vector/fragments/RoomsFragment.java
@@ -25,10 +25,8 @@ import android.support.v7.widget.DividerItemDecoration;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.text.TextUtils;
-import android.view.LayoutInflater;
 import android.view.MotionEvent;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.Filter;
 import android.widget.Spinner;
@@ -39,7 +37,6 @@ import org.matrix.androidsdk.data.RoomPreviewData;
 import org.matrix.androidsdk.rest.callback.ApiCallback;
 import org.matrix.androidsdk.rest.client.EventsRestClient;
 import org.matrix.androidsdk.rest.model.MatrixError;
-import org.matrix.androidsdk.rest.model.RoomMember;
 import org.matrix.androidsdk.rest.model.publicroom.PublicRoom;
 import org.matrix.androidsdk.util.Log;
 
@@ -490,7 +487,7 @@ public class RoomsFragment extends AbsHomeFragment implements AbsHomeFragment.On
                 mCurrentFilter, new ApiCallback<List<PublicRoom>>() {
                     @Override
                     public void onSuccess(List<PublicRoom> publicRooms) {
-                        if (null != getActivity()) {
+                        if (isAdded()) {
                             mAdapter.setNoMorePublicRooms(publicRooms.size() < PublicRoomsManager.PUBLIC_ROOMS_LIMIT);
                             mAdapter.setPublicRooms(publicRooms);
                             addPublicRoomsListener();
@@ -516,7 +513,7 @@ public class RoomsFragment extends AbsHomeFragment implements AbsHomeFragment.On
                     }
 
                     private void onError(String message) {
-                        if (null != getActivity()) {
+                        if (isAdded()) {
                             Log.e(LOG_TAG, "## startPublicRoomsSearch() failed " + message);
                             Toast.makeText(getActivity(), message, Toast.LENGTH_SHORT).show();
                             hidePublicRoomsLoadingView();
@@ -551,7 +548,7 @@ public class RoomsFragment extends AbsHomeFragment implements AbsHomeFragment.On
         boolean isForwarding = PublicRoomsManager.getInstance().forwardPaginate(new ApiCallback<List<PublicRoom>>() {
             @Override
             public void onSuccess(final List<PublicRoom> publicRooms) {
-                if (null != getActivity()) {
+                if (isAdded()) {
                     // unplug the scroll listener if there is no more data to find
                     if (!PublicRoomsManager.getInstance().hasMoreResults()) {
                         mAdapter.setNoMorePublicRooms(true);
@@ -564,12 +561,12 @@ public class RoomsFragment extends AbsHomeFragment implements AbsHomeFragment.On
             }
 
             private void onError(String message) {
-                if (null != getActivity()) {
+                if (isAdded()) {
                     Log.e(LOG_TAG, "## forwardPaginate() failed " + message);
                     Toast.makeText(getActivity(), message, Toast.LENGTH_SHORT).show();
-                }
 
-                hidePublicRoomsLoadingView();
+                    hidePublicRoomsLoadingView();
+                }
             }
 
             @Override
@@ -606,7 +603,7 @@ public class RoomsFragment extends AbsHomeFragment implements AbsHomeFragment.On
      * Remove the public rooms listener
      */
     private void removePublicRoomsListener() {
-        mRecycler.removeOnScrollListener(null);
+        mRecycler.removeOnScrollListener(mPublicRoomScrollListener);
     }
 
     /*