+++ CHANGES.rst
@@ -11,6 +11,7 @@ Improvements:
  - Support Split-screen mode (#1832)
  - Enable auto focus when taking picture with the camera (#2831)
  - Better wording in notification for video call (#1421)
+ - Improve widget banner (#2129)
 
 Other changes:
  - Restore the crash report dialog after a crash